refactored handler to module
